sync code

+++ 2015-CaroloCup/sources/OpenDaVINCI-msv/apps/2015/proxy/include/Proxy.h
@@ -77,7 +77,7 @@ namespace msv {
 
         public:
             struct vehicleControl {
-                    		double speed;
+                    		int speed;
                     		double steeringAngle;
                     		bool leftFlash;
                     		bool rightFlash;
